requirements.txt
fastapi==0.104.1
celery==5.3.4
redis==5.0.1
pydantic==2.5.2
openai==1.6.1
python-multipart==0.0.6
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
PyMuPDF==1.23.8